+++ src/component/PostAll/index.module.css
@@ -1,0 +1,4 @@
+.PostAll{
+border-bottom:2px gray solid;
+padding-bottom: 10px;
+}
