+++ src/component/PostImage/index.module.css
@@ -1,4 +1,3 @@
 .Image{
-    background-color: aqua;
     border-radius: 2rem;
 }
